requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
boto3==1.33.11
botocore==1.33.11
cachetools==5.3.2
carla==0.9.14
certifi==2023.7.22
charset-normalizer==3.2.0
clearml==1.13.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.5
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.1.1
croniter==1.4.1
cycler==0.11.0
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.5.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.103.1
fastjsonschema==2.19.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
furl==2.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.24.0
google-auth-oauthlib==1.0.0
grpcio==1.59.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.0
inquirer==3.1.3
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==17.0.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.8.1.78
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.4.0
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
s3transfer==0.8.2
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
sentry-sdk==1.39.2
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stable-baselines3==2.2.1
stack-data==0.6.3
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
torch==2.0.1
torch-scatter==2.1.2+pt20cu117
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.10.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
wandb==0.16.2
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
